pyproject.toml
[tool.poetry]
name = "ens-data"
version = "0.1.0"
description = ""
authors = ["Sergey Nesterov <sergioreklamplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
web3 = "^6.19.0"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
ens-normalize = "^3.0.7"
pysha3 = "^1.0.2"
pytest-django = "^4.9.0"
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py", "*/tests.py"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

exclude = '''
(
  /(
      .eggs         # exclude a few common directories in the
    | .git          # root of the project
    | .mypy_cache
    | .venv
    | static
    | /*/migrations
    | /*/*/migrations
  )/
  | manage.py
)
'''

[tool.isort]
profile = "black"
